Relocate LegacyAddOperationTransformation to ProtocolResourceDefinition
